Refactor evaluations module packages
