+++ evaluations/src/main/kotlin/com/gdavidpb/tuindice/evaluations/data/repository/evaluation/source/PreferencesDataSource.kt
@@ -1,8 +1,8 @@
-package com.gdavidpb.tuindice.evaluations.data.preferences
+package com.gdavidpb.tuindice.evaluations.data.repository.evaluation.source
 
 import android.content.SharedPreferences
 import androidx.core.content.edit
-import com.gdavidpb.tuindice.evaluations.data.evaluation.source.SettingsDataSource
+import com.gdavidpb.tuindice.evaluations.data.repository.evaluation.SettingsDataSource
 import com.gdavidpb.tuindice.evaluations.utils.CooldownTimes
 import com.gdavidpb.tuindice.evaluations.utils.PreferencesKeys
