+++ evaluations/src/main/kotlin/com/gdavidpb/tuindice/evaluations/di/EvaluationsModule.kt
@@ -2,18 +2,18 @@ package com.gdavidpb.tuindice.evaluations.di
 
 import com.gdavidpb.tuindice.base.BuildConfig
 import com.gdavidpb.tuindice.base.utils.extension.create
-import com.gdavidpb.tuindice.evaluations.data.api.ApiDataSource
-import com.gdavidpb.tuindice.evaluations.data.api.EvaluationsApi
-import com.gdavidpb.tuindice.evaluations.data.api.parser.EvaluationAddParser
-import com.gdavidpb.tuindice.evaluations.data.api.parser.EvaluationRemoveParser
-import com.gdavidpb.tuindice.evaluations.data.api.parser.EvaluationUpdateParser
-import com.gdavidpb.tuindice.evaluations.data.evaluation.EvaluationDataRepository
-import com.gdavidpb.tuindice.evaluations.data.evaluation.source.LocalDataSource
-import com.gdavidpb.tuindice.evaluations.data.evaluation.source.RemoteDataSource
-import com.gdavidpb.tuindice.evaluations.data.evaluation.source.SettingsDataSource
-import com.gdavidpb.tuindice.evaluations.data.preferences.PreferencesDataSource
-import com.gdavidpb.tuindice.evaluations.data.resolution.EvaluationResolutionHandler
-import com.gdavidpb.tuindice.evaluations.data.room.RoomDataSource
+import com.gdavidpb.tuindice.evaluations.data.repository.evaluation.source.EvaluationsApiDataSource
+import com.gdavidpb.tuindice.evaluations.data.repository.evaluation.EvaluationsApi
+import com.gdavidpb.tuindice.evaluations.data.repository.evaluation.source.api.parser.EvaluationAddParser
+import com.gdavidpb.tuindice.evaluations.data.repository.evaluation.source.api.parser.EvaluationRemoveParser
+import com.gdavidpb.tuindice.evaluations.data.repository.evaluation.source.api.parser.EvaluationUpdateParser
+import com.gdavidpb.tuindice.evaluations.data.repository.evaluation.EvaluationDataRepository
+import com.gdavidpb.tuindice.evaluations.data.repository.evaluation.LocalDataSource
+import com.gdavidpb.tuindice.evaluations.data.repository.evaluation.RemoteDataSource
+import com.gdavidpb.tuindice.evaluations.data.repository.evaluation.SettingsDataSource
+import com.gdavidpb.tuindice.evaluations.data.repository.evaluation.source.PreferencesDataSource
+import com.gdavidpb.tuindice.evaluations.data.repository.evaluation.source.database.resolution.EvaluationResolutionHandler
+import com.gdavidpb.tuindice.evaluations.data.repository.evaluation.source.RoomDataSource
 import com.gdavidpb.tuindice.evaluations.domain.repository.EvaluationRepository
 import com.gdavidpb.tuindice.evaluations.domain.usecase.AddEvaluationUseCase
 import com.gdavidpb.tuindice.evaluations.domain.usecase.GetAvailableSubjectsUseCase
@@ -106,7 +106,7 @@ val evaluationsModule = module {
 
 	/* Data sources */
 
-	factoryOf(::ApiDataSource) { bind<RemoteDataSource>() }
+	factoryOf(::EvaluationsApiDataSource) { bind<RemoteDataSource>() }
 	factoryOf(::RoomDataSource) { bind<LocalDataSource>() }
 	factoryOf(::PreferencesDataSource) { bind<SettingsDataSource>() }
